requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.10.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
EZSheets==2024.8.9
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.27.2
idna==3.10
IMAPClient==3.0.1
iniconfig==2.0.0
lxml==4.8.0
MouseInfo==0.1.3
oauthlib==3.2.0
ollama==0.4.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.4.8
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pydantic==2.10.4
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyInputPlus==0.2.12
PyMsgBox==1.0.9
pyparsing==2.4.7
pypdf==5.1.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySimpleValidate==0.2.12
PySocks==1.7.1
pytest==8.3.3
python-docx==1.1.2
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pyzmail36==1.0.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.1.0
soupsieve==2.3.1
stdiomask==0.0.6
tomli==1.2.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.0.0
